client: Refactor directory structure
This is still in progress:

* client/ contains application-specific components and other code related to
  them, possibly divided to subdirectories (like "reducers").
* client/shared/ contains generic components.
* client/shared/logic/ contains reusable logic used by generic components.
