+++ client/index.js
@@ -8,7 +8,7 @@ const React = require('react');
 const { render } = require('react-dom');
 const { Provider } = require('react-redux');
 //const { addAllStyles } = require('./styles');
-const App = require('./components/App');
+const App = require('./App');
 const { createLogger } = require('redux-logger');
 const thunk = require('redux-thunk').default;
 const reducer = require('./reducers');
