+++ client/shared/EditPanel.js
